+++ Automate/Framework/Commands/Summary/GroupStats.cs
@@ -1,4 +1,5 @@
 using System.Linq;
+using Microsoft.Xna.Framework;
 
 namespace Pathoschild.Stardew.Automate.Framework.Commands.Summary
 {
@@ -28,9 +29,13 @@ namespace Pathoschild.Stardew.Automate.Framework.Commands.Summary
         /// <param name="machineGroup">The machine group to analyze.</param>
         public GroupStats(IMachineGroup machineGroup)
         {
-            this.Name = machineGroup.IsJunimoGroup
-                ? "Distributed group"
-                : $"Group at ({machineGroup.Tiles[0].X}, {machineGroup.Tiles[0].Y})";
+            if (machineGroup.IsJunimoGroup)
+                this.Name = "Distributed group";
+            else
+            {
+                Vector2 tile = machineGroup.Tiles.First();
+                this.Name = $"Group at ({tile.X}, {tile.Y})";
+            }
 
             this.IsJunimoGroup = machineGroup.IsJunimoGroup;