+++ ContentPatcher/Framework/Migrations/Migration_1_18.cs
@@ -8,7 +8,7 @@ using StardewModdingAPI;
 
 namespace ContentPatcher.Framework.Migrations
 {
-    /// <summary>Validates patches for format version 1.18.</summary>
+    /// <summary>Migrates patches to format version 1.18.</summary>
     [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Named for clarity.")]
     internal class Migration_1_18 : BaseMigration
     {