+++ ContentPatcher/Framework/Migrations/Migration_1_23.cs
@@ -10,7 +10,7 @@ using StardewModdingAPI;
 
 namespace ContentPatcher.Framework.Migrations
 {
-    /// <summary>Validates patches for format version 1.23.</summary>
+    /// <summary>Migrates patches to format version 1.23.</summary>
     [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Named for clarity.")]
     internal class Migration_1_23 : BaseMigration
     {