+++ ContentPatcher/Framework/Migrations/Migration_1_27.cs
@@ -3,15 +3,15 @@ using StardewModdingAPI;
 
 namespace ContentPatcher.Framework.Migrations
 {
-    /// <summary>Migrates patches to format version 1.0.</summary>
+    /// <summary>Migrates patches to format version 1.27.</summary>
     [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Named for clarity.")]
-    internal class Migration_1_0 : BaseMigration
+    internal class Migration_1_27 : BaseMigration
     {
         /*********
         ** Public methods
         *********/
         /// <summary>Construct an instance.</summary>
-        public Migration_1_0()
-            : base(new SemanticVersion(1, 0, 0)) { }
+        public Migration_1_27()
+            : base(new SemanticVersion(1, 27, 0)) { }
     }
 }