+++ ContentPatcher/Framework/Migrations/Migration_1_3.cs
@@ -6,7 +6,7 @@ using StardewModdingAPI;
 
 namespace ContentPatcher.Framework.Migrations
 {
-    /// <summary>Migrate patches to format version 1.3.</summary>
+    /// <summary>Migrates patches to format version 1.3.</summary>
     [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Named for clarity.")]
     internal class Migration_1_3 : BaseMigration
     {