+++ ContentPatcher/Framework/Migrations/Migration_1_4.cs
@@ -5,7 +5,7 @@ using StardewModdingAPI;
 
 namespace ContentPatcher.Framework.Migrations
 {
-    /// <summary>Migrate patches to format version 1.4.</summary>
+    /// <summary>Migrates patches to format version 1.4.</summary>
     [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Named for clarity.")]
     internal class Migration_1_4 : BaseMigration
     {