+++ ContentPatcher/Framework/Migrations/Migration_1_8.cs
@@ -8,7 +8,7 @@ using StardewModdingAPI;
 
 namespace ContentPatcher.Framework.Migrations
 {
-    /// <summary>Migrate patches to format version 1.8.</summary>
+    /// <summary>Migrates patches to format version 1.8.</summary>
     [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Named for clarity.")]
     internal class Migration_1_8 : BaseMigration
     {